+++ lib/src/widgets/shadow_button.dart
@@ -57,7 +57,7 @@ class ShadowButton extends StatelessWidget {
       decoration: BoxDecoration(
         borderRadius: borderRadius,
         color: Colors.transparent,
-        boxShadow: [if (boxShadow != null) boxShadow!],
+        boxShadow: [?boxShadow],
       ),
       child: Material(
         shape: RoundedRectangleBorder(borderRadius: borderRadius),